requirements.txt
django~=4.2
black~=23.3.0
natasha~=1.5.0
psycopg~=3.1.8
django-filter~=23.2